Added an InputHandler and an Input group system
Added an Input Handler class that can create InputGroup's which are used to handle a group of inputs instead of all inputs being handled at once.
This is usefull for whenever there should only be a few inputs for a menu whilst when moving about the 'whatever we are gonna call the world', we can have more inputs.